+++ src/utils/time.ts
@@ -1,0 +1,5 @@
+export const fmt = (total: number) => {
+  const m = Math.floor(total / 60);
+  const s = total % 60;
+  return `${String(m).padStart(2, "0")}:${String(s).padStart(2, "0")}`;
+};
